+++ frontend/src/components/common/footer.js
