+++ frontend/src/components/common/navBar.js
