+++ frontend/src/components/display/hours/hoursDataView.js
