+++ frontend/src/pages/NotificationPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Notifications from "../components/elements/Notifications";
+import Notifications from "../components/common/Notifications";
 import Container from "@material-ui/core/Container";
 
 class NotificationPage extends React.Component {
